+++ src/sklearndf/transformation/wrapper/_wrapper.py
@@ -64,7 +64,7 @@ T_Transformer = TypeVar("T_Transformer", bound=TransformerMixin)
 # onwards.
 # Once we drop support for sklearn 0.21, _BaseImputer can be used instead.
 # The following TypeVar helps to annotate availability of "add_indicator" and
-# "missing_values" attributes on an imputer instance for ImputerWrapperDF below
+# "missing_values" attributes on an imputer instance for ImputerWrapperDF below.
 
 # noinspection PyProtectedMember
 from sklearn.impute._iterative import IterativeImputer
@@ -268,10 +268,10 @@ class ColumnTransformerWrapperDF(
     :class:`.TransformerDF`.
     """
 
-    __DROP = "drop"
-    __PASSTHROUGH = "passthrough"
+    DROP = "drop"
+    PASSTHROUGH = "passthrough"
 
-    __SPECIAL_TRANSFORMERS = (__DROP, __PASSTHROUGH)
+    __SPECIAL_TRANSFORMERS = (DROP, PASSTHROUGH)
 
     def _validate_delegate_estimator(self) -> None:
         column_transformer: ColumnTransformer = self.native_estimator
@@ -326,7 +326,7 @@ class ColumnTransformerWrapperDF(
             input_column_names: npt.NDArray[Any]
             output_column_names: npt.NDArray[Any]
 
-            if df_transformer == ColumnTransformerWrapperDF.__PASSTHROUGH:
+            if df_transformer == ColumnTransformerWrapperDF.PASSTHROUGH:
                 # we may get positional indices for columns selected by the
                 # 'passthrough' transformer, and in that case so need to look up the
                 # associated column names
@@ -368,7 +368,7 @@ class ColumnTransformerWrapperDF(
                 )
                 if (
                     len(columns) > 0
-                    and df_transformer != ColumnTransformerWrapperDF.__DROP
+                    and df_transformer != ColumnTransformerWrapperDF.DROP
                 )
             ]
         )
@@ -500,8 +500,8 @@ class OneHotEncoderWrapperDF(TransformerWrapperDF[OneHotEncoder], metaclass=ABCM
     def _get_features_original(self) -> pd.Series:
         # Return the series mapping output column names to original column names.
         #
-        # Remove 1st category column if argument drop == 'first'
-        # Remove 1st category column only of binary features if arg drop == 'if_binary'
+        # Remove 1st category column if argument drop == 'first'.
+        # Remove 1st category column only of binary features if arg drop == 'if_binary'.
 
         native_estimator: OneHotEncoder = self.native_estimator
         feature_names_in: pd.Index = self.feature_names_in_